requirements.txt
flask
flask-cors
phidata
openai
duckduckgo-search 
sqlalchemy
